backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
celery
redis
PyMuPDF
gTTS
moviepy
Pillow
python-dotenv
ffmpeg-python
moviepy